backend/offer-agent/src copy/requirements.txt
aiofiles==24.1.0
aiohttp==3.12.12
aiomqtt==2.4.0
emails==0.6
fastapi==0.115.14
google_api_python_client==2.171.0
google_auth_oauthlib==1.2.2
httpx==0.28.1
Jinja2==3.1.6
numpy==2.3.1
openai==1.93.0
openpyxl==3.1.5
pandas==2.3.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_settings==2.10.1
pyodbc==5.2.0
qdrant_client==1.14.3
redis==6.2.0
reportlab==4.4.1
SQLAlchemy==2.0.41
structlog==25.4.0
uvicorn==0.35.0
mistralai